requirements.txt
ipython==7.16.1
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
## The following requirements were added by pip freeze:
APScheduler==3.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2020.4.5.2
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
importlib-resources==5.2.2
ipython-genutils==0.2.0
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
six==1.15.0
tornado==6.1
traitlets==4.3.3
tzlocal==3.0
wcwidth==0.2.4
zipp==3.5.0
